// packages/frontend/src/typings/requests/EditCustomerRequest.ts
export interface EditCustomerRequest {
  name?: string;
  age?: number;
  email?: string;
  phone?: string;
  favoriteGenres?: string[];
  address?: string;
  complement?: string;
  number?: string;
  neighborhood?: string;
  city?: string;
  zipCode?: string;
  state?: string;
  country?: string;
}